Refactor oxc preprocessing helpers
